+++ bin/templates/components/entityeventhandler/template.js
@@ -1,5 +1,7 @@
 'use strict';
 
+// Documentation for writing entity event handlers: https://github.com/oracle/bots-node-sdk/blob/master/ENTITY_EVENT_HANDLER.md
+
 // You can use your favorite http client package to make REST calls, however, the node fetch API is pre-installed with the bots-node-sdk.
 // Documentation can be found at https://www.npmjs.com/package/node-fetch
 // Un-comment the next line if you want to make REST calls using node-fetch.