+++ ts/lib/dataquery/index.ts
@@ -1,0 +1,3 @@
+export * from './dataQueryContext';
+export * from './dataQueryTypes'; // type definitions only
+