+++ ts/lib/index.ts
@@ -1,4 +1,5 @@
 export * from './component';
 export * from './entity';
+export * from './dataquery';
 export * from './message';
 export { InvocationCallback } from '../common/definitions';